added set image method. refactoring.
