+++ app/build.gradle
@@ -41,7 +41,7 @@ dependencies {
     implementation "androidx.recyclerview:recyclerview:1.1.0"
     implementation "androidx.annotation:annotation:1.1.0"
     implementation "com.google.android.material:material:1.1.0"
-
+    implementation 'com.squareup.picasso:picasso:2.71828'
     implementation "com.mikepenz:materialdrawer-nav:7.0.0"
     implementation 'de.hdodenhof:circleimageview:3.1.0'
     api 'com.theartofdev.edmodo:android-image-cropper:2.8.0'
