+++ app/src/main/java/com/mornelytom/telegram/MainActivity.kt
@@ -28,10 +28,6 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         mBinding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(mBinding.root)
-    }
-
-    override fun onStart() {
-        super.onStart()
         APP_ACTIVITY = this
         initFields()
         initFunc()
@@ -60,39 +56,6 @@ class MainActivity : AppCompatActivity() {
             .addListenerForSingleValueEvent(AppValueEventListener {
                 USER = it.getValue(User::class.java) ?: User()
             })
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == CropImage.CROP_IMAGE_ACTIVITY_REQUEST_CODE
-            && resultCode == Activity.RESULT_OK && data != null
-        ) {
-            val uri = CropImage.getActivityResult(data).uri
-            val path = REF_STORAGE_ROOT.child(FOLDER_PROFILE_IMAGE).child(UID)
-            path.putFile(uri).addOnCompleteListener { putTask ->
-                if (putTask.isSuccessful) {
-                    path.downloadUrl.addOnCompleteListener { downloadTask ->
-                        if (downloadTask.isSuccessful) {
-                            val photoUrl = downloadTask.result.toString()
-                            REF_DATABASE_ROOT.child(NODE_USERS).child(UID).child(USER_PHOTO_URL)
-                                .setValue(photoUrl).addOnCompleteListener { databaseTask ->
-                                    if (databaseTask.isSuccessful) {
-                                        showToast(getString(R.string.toast_data_update))
-                                        USER.photoUrl = photoUrl
-                                    }
-                                }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fun hideKeyboard() {
-        val imm: InputMethodManager =
-            getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-        imm.hideSoftInputFromWindow(window.decorView.windowToken, 0)
-
     }
 
 }
