+++ app/src/main/java/com/mornelytom/telegram/ui/fragments/SettingsFragment.kt
@@ -1,11 +1,14 @@
 package com.mornelytom.telegram.ui.fragments
 
+import android.app.Activity
+import android.content.Intent
 import android.view.Menu
 import android.view.MenuInflater
 import android.view.MenuItem
 import com.mornelytom.telegram.R
 import com.mornelytom.telegram.activities.AuthorizationActivity
 import com.mornelytom.telegram.utilits.*
+import com.squareup.picasso.Picasso
 import com.theartofdev.edmodo.cropper.CropImage
 import com.theartofdev.edmodo.cropper.CropImageView
 import kotlinx.android.synthetic.main.fragment_settings.*
@@ -49,6 +52,37 @@ class SettingsFragment : BaseFragment(R.layout.fragment_settings) {
             .setAspectRatio(1, 1)
             .setRequestedSize(600, 600)
             .setCropShape(CropImageView.CropShape.OVAL)
-            .start(APP_ACTIVITY)
+            .start(APP_ACTIVITY, this)
+    }
+
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (requestCode == CropImage.CROP_IMAGE_ACTIVITY_REQUEST_CODE
+            && resultCode == Activity.RESULT_OK && data != null
+        ) {
+            // selected image
+            val uri = CropImage.getActivityResult(data).uri
+            // upload image to storage
+            val path = REF_STORAGE_ROOT.child(FOLDER_PROFILE_IMAGE).child(UID)
+            path.putFile(uri).addOnCompleteListener { putTask ->
+                if (putTask.isSuccessful) {
+                    // download image from storage
+                    path.downloadUrl.addOnCompleteListener { downloadTask ->
+                        if (downloadTask.isSuccessful) {
+                            val photoUrl = downloadTask.result.toString()
+                            REF_DATABASE_ROOT.child(NODE_USERS).child(UID).child(USER_PHOTO_URL)
+                                .setValue(photoUrl).addOnCompleteListener { databaseTask ->
+                                    if (databaseTask.isSuccessful) {
+                                        settings_user_photo.downloadAndSetImage(photoUrl)
+                                        showToast(getString(R.string.toast_data_update))
+                                        USER.photoUrl = photoUrl
+                                    }
+                                }
+                        }
+                    }
+                }
+            }
+        }
     }
 }
