+++ app/src/main/java/com/mornelytom/telegram/utilits/func.kt
@@ -1,10 +1,15 @@
 package com.mornelytom.telegram.utilits
 
+import android.content.Context
 import android.content.Intent
+import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.mornelytom.telegram.R
+import com.squareup.picasso.Picasso
+import de.hdodenhof.circleimageview.CircleImageView
+import kotlinx.android.synthetic.main.fragment_settings.*
 
 fun showToast(message: String) {
     Toast.makeText(APP_ACTIVITY, message, Toast.LENGTH_SHORT).show()
@@ -44,4 +49,19 @@ fun Fragment.replaceFragment(fragment: Fragment) {
             fragment
         )
         ?.commit()
+}
+
+
+fun hideKeyboard() {
+    val imm: InputMethodManager =
+        APP_ACTIVITY.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+    imm.hideSoftInputFromWindow(APP_ACTIVITY.window.decorView.windowToken, 0)
+
+}
+
+fun CircleImageView.downloadAndSetImage(url: String) {
+    Picasso.get()
+        .load(url)
+        .placeholder(R.drawable.default_user)
+        .into(this)
 }
